cybersecurity/trainingPortal/migrations/0006_auto_20150124_1625.py
# -*- coding: utf-8 -*-
from __future__ import unicode_literals

from django.db import models, migrations
import tinymce.models


class Migration(migrations.Migration):

    dependencies = [
        ('trainingPortal', '0005_page_content'),
    ]

    operations = [
        migrations.AlterField(
            model_name='page',
            name='content',
            field=tinymce.models.HTMLField(default=b''),
            preserve_default=True,
        ),
    ]
