cybersecurity/requirements.txt
Codeville==0.8.0
Django==1.7.2
Mako==0.7.2
MarkupSafe==0.15
PIL==1.1.7
Pillow==2.7.0
Pygments==1.5
Pyrex==0.9.9
Twisted==12.3.0
Unidecode==0.04.17
argparse==1.3.0
bsddb3==5.3.0
bzr==2.6.0
bzr-fastimport==0.13.0.final.0
chardet==2.0.1
cvs2svn==2.4.0
dblatex==0.3.4
dj-database-url==0.3.0
django-allauth==0.19.0
django-bootstrap==0.2.4
django-bootstrap-form==3.1
django-ckeditor-updated==4.4.4
django-classy-tags==0.5.2
django-cms==3.0.9
django-filer==0.9.8
django-forms-bootstrap==3.0.1
django-mptt==0.6.1
django-polymorphic==0.6.1
django-reversion==1.8.5
django-sekizai==0.8.1
django-tinymce==1.5.3
django-wysiwyg==0.7.0
djangocms-admin-style==0.2.5
djangocms-file==0.1
djangocms-flash==0.1
djangocms-googlemap==0.2
djangocms-inherit==0.1
djangocms-installer==0.7.1
djangocms-link==1.5
djangocms-picture==0.1
djangocms-teaser==0.1
djangocms-text-ckeditor==2.4.3
djangocms-video==0.1
easy-thumbnails==2.2
fastimport==0.9.2
feedparser==5.0.1
gdata==2.0.18
getmail==4.36.0
googlecl==0.9.14
html5lib==0.999
htmltmpl==1.22
logilab-astng==0.23.1
logilab-common==0.58.0
mercurial==3.0.1
mingc==0.4.5
netsnmp-python==1.0a1
numpy==1.7.2
oauthlib==0.7.2
offlineimap==6.5.4
paramiko==1.14.0
planet==2.0
pyOpenSSL==0.13.1
pyasn1==0.1.6
pycrypto==2.6
pygobject==3.10.2
pylint==0.25.1
python-openid==2.2.5
pytz==2014.10
pyxdg==0.25
rdiff-backup==1.2.8
requests==2.5.1
requests-oauthlib==0.4.2
six==1.9.0
spambayes==1.0.4
stgit==0.17.1-dirty
tailor==0.9.35
urlgrabber==3.1.0
virtualenv==1.11.4
wsgiref==0.1.2
zope.interface==4.0.3
